+++ src/librpbase/img/RpGdiplusBackend.hpp
@@ -62,15 +62,6 @@ class RpGdiplusBackend : public rp_image_backend
 		 */
 		RpGdiplusBackend(int width, int height, rp_image::Format format);
 
-		/**
-		 * Create an RpGdiplusBackend using the specified Gdiplus::Bitmap.
-		 *
-		 * NOTE: This RpGdiplusBackend will take ownership of the Gdiplus::Bitmap.
-		 *
-		 * @param pGdipBmp Gdiplus::Bitmap.
-		 */
-		explicit RpGdiplusBackend(Gdiplus::Bitmap *pGdipBmp);
-
 		virtual ~RpGdiplusBackend();
 
 	private:
@@ -218,8 +209,12 @@ class RpGdiplusBackend : public rp_image_backend
 
 		// BitmapData for locking.
 		bool m_isLocked;
+		uint8_t m_bytesppShift;	// bytespp shift value.
 		Gdiplus::PixelFormat m_gdipFmt;
 		Gdiplus::BitmapData m_gdipBmpData;
+
+		// Allocated image buffer for Bitmap::LockBits().
+		void *m_pImgBuf;
 
 		// Color palette.
 		// Pointer to Entries[0] is used for rp_image_backend::palette.